chore: move configuration to config folder
