+++ packages/credential-provider-imds/src/config/Endpoint.ts
