+++ packages/credential-provider-imds/src/config/EndpointConfigOptions.spec.ts
