+++ packages/credential-provider-imds/src/config/EndpointConfigOptions.ts
