+++ packages/credential-provider-imds/src/config/EndpointMode.ts
