+++ packages/credential-provider-imds/src/config/EndpointModeConfigOptions.spec.ts
