+++ packages/credential-provider-imds/src/config/EndpointModeConfigOptions.ts
